fix(timings): attribute startup work to completed stages
Report runtime milestones at their actual completion boundaries so guest negotiation owns the guest wait instead of compute launch.

Carry nanosecond runner observations and publish fractional milliseconds. This preserves the existing unit while retaining sub-millisecond stages that whole-millisecond projection rounded to zero.

+++ tests/scenario/stress/api.go
@@ -250,7 +250,7 @@ func (driver *stressDriver) collectBootTiming(ctx context.Context, operationID s
 		for _, stage := range boot.Stages {
 			driver.bootStages[stage.Stage] = append(
 				driver.bootStages[stage.Stage],
-				time.Duration(stage.ElapsedMilliseconds)*time.Millisecond,
+				time.Duration(stage.ElapsedMilliseconds*float64(time.Millisecond)),
 			)
 		}
 	}
